+++ docs/changes.rst
@@ -108,6 +108,8 @@ Compatibility notes:
 - removed --bypass-lock option
 - removed --remote-path option (use the BORG_REMOTE_PATH environment variable)
 - removed --rsh option (use the BORG_RSH environment variable)
+- removed --upload-ratelimit and --upload-buffer options (they only ever
+  affected uploads to a borg 1.x repository via ssh://)
 - removed borg config command (only worked locally anyway)
 - compact command now requires access to the borg key if the repo is encrypted
   or authenticated